[src] wallet: move responsibility of chain check out of deser_wallet()
A wallet can be deserialized as long as libccoin supports and knows
about the chain associated with it's network magic.

Because the notion of a "current chain" is specific to picocoin, move
the responsibility of the check into the picocoin-specific load_wallet()
function.  Also, simplify the memcmp(netmagic) to a simple pointer
comparison into the chain_info table.