[test] wallet: fixup temporary wallet allocation and asserts
Because the temporary wallet was being allocated on the stack, it's
members (including the parr keys member) were uninitialized, and
therefore adding a new key segfaulted.

Fix this up by allocating a new temporary wallet using wallet_new().

Also, fixup the assertion logic so it ensures deser_wallet() _succeeds_
instead of fails (a logic error in the initial commit).

With these modifications, the serialization tests pass, so re-add them.